chore: refactor generator.ts
